Added Subscription Alert to users
This creates an alert on every login, if your subscription is ending in a
week, or notifies a user once if they have not been alerted that their
subscription is ending in a weeke naturally by a stored cookie value.

This needs to be tested with the backend, but Julian currently holds the
database. which I cannot access from home