Go To Page Bug Fixed, Error / Alert Handler
Fixed the Go To page bug by adding an ng-disable on the Go button,
And Implemented an error alert handler Julian and I talked about
in person since It was needed to be done in a later issue ( #31 ,
and #34  ).

Everything Looked good, but the backend was acting up, so please test
this.

closes #37